[Python] Rename ngraph APIs
