+++ python/examples/mnist_mlp.py
@@ -14,12 +14,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ******************************************************************************
-from ngraph import Type, Function
-from ngraph import Node
-from ngraph.op import Parameter, Maximum, Reshape, Dot, Broadcast
-from ngraph.op import Constant, Exp, Log, Sum
-from ngraph.op import Greater, Convert, Reduce
-from ngraph.op import OneHot
+from ngraph_bind import Type, Function
+from ngraph_bind import Node
+from ngraph_bind.op import Parameter, Maximum, Reshape, Dot, Broadcast
+from ngraph_bind.op import Constant, Exp, Log, Sum
+from ngraph_bind.op import Greater, Convert, Reduce
+from ngraph_bind.op import OneHot
 
 from typing import List, Dict, Set
