+++ python/ngraph/__init__.py
@@ -1,5 +1,5 @@
 # ******************************************************************************
-# Copyright 2017-2018 Intel Corporation
+# Copyright 2018 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,36 +13,43 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ******************************************************************************
-"""
-Package: ngraph
-Low level wrappers for the nGraph c++ api.
-"""
+"""ngraph module namespace, exposing factory functions for all ops and other classes."""
 
-# flake8: noqa
+from ngraph.ops import absolute
+from ngraph.ops import absolute as abs
+from ngraph.ops import add
+from ngraph.ops import avg_pool
+from ngraph.ops import broadcast
+from ngraph.ops import ceiling
+from ngraph.ops import ceiling as ceil
+from ngraph.ops import constant
+from ngraph.ops import convert
+from ngraph.ops import convolution
+from ngraph.ops import divide
+from ngraph.ops import dot
+from ngraph.ops import equal
+from ngraph.ops import exp
+from ngraph.ops import floor
+from ngraph.ops import greater
+from ngraph.ops import greater_eq
+from ngraph.ops import log
+from ngraph.ops import less
+from ngraph.ops import less_eq
+from ngraph.ops import logical_not
+from ngraph.ops import max
+from ngraph.ops import maximum
+from ngraph.ops import max_pool
+from ngraph.ops import min
+from ngraph.ops import minimum
+from ngraph.ops import multiply
+from ngraph.ops import negative
+from ngraph.ops import not_equal
+from ngraph.ops import parameter
+from ngraph.ops import prod
+from ngraph.ops import reshape
+from ngraph.ops import sqrt
+from ngraph.ops import subtract
+from ngraph.ops import sum
+from ngraph.ops import tanh
 
-import sys
-import six
-
-# workaround to load the libngraph.so with RTLD_GLOBAL
-if six.PY3:
-    import os
-    flags = os.RTLD_NOW | os.RTLD_GLOBAL
-else:
-    import ctypes
-    flags = sys.getdlopenflags() | ctypes.RTLD_GLOBAL
-sys.setdlopenflags(flags)
-
-from _pyngraph import Function
-from _pyngraph import Node
-from _pyngraph import NodeVector
-from _pyngraph import Type
-from _pyngraph import TensorViewType
-from _pyngraph import Shape
-from _pyngraph import Strides
-from _pyngraph import CoordinateDiff
-from _pyngraph import AxisSet
-from _pyngraph import AxisVector
-from _pyngraph import Coordinate
-
-from _pyngraph import serialize
-from _pyngraph import util
+from ngraph.runtime import runtime
