+++ python/ngraph/exceptions.py
