+++ python/ngraph/utils/broadcasting.py
@@ -16,10 +16,10 @@
 import logging
 from typing import Optional, List
 
-import ngraph_api as ng
+import ngraph as ng
 
-from ngraph import AxisSet, Node
-from ngraph_api.utils.types import TensorShape, get_dtype, make_constant_node, NodeInput
+from ngraph_bind import AxisSet, Node
+from ngraph.utils.types import TensorShape, get_dtype, make_constant_node, NodeInput
 
 log = logging.getLogger(__file__)
