+++ python/setup.py.in
@@ -255,14 +255,14 @@ setup(
     description='Python wrapper for ngraph',
     long_description='',
     ext_modules=ext_modules,
-    package_dir={'ngraph': PYNGRAPH_SOURCE_DIR + "/ngraph",
-                 'ngraph.op': PYNGRAPH_SOURCE_DIR + "/ngraph/op",
-                 'ngraph.op.util': PYNGRAPH_SOURCE_DIR + "/ngraph/op/util",
-                 'ngraph.passes': PYNGRAPH_SOURCE_DIR + "/ngraph/passes",
-                 'ngraph.runtime': PYNGRAPH_SOURCE_DIR + "/ngraph/runtime",
-                 'ngraph_api': PYNGRAPH_SOURCE_DIR + "/ngraph_api",
-                 'ngraph_api.utils': PYNGRAPH_SOURCE_DIR + "/ngraph_api/utils"},
-    packages = ['ngraph', 'ngraph.op', 'ngraph.op.util', 'ngraph.passes', 'ngraph.runtime', 'ngraph_api', 'ngraph_api.utils'],
+    package_dir={'ngraph_bind': PYNGRAPH_SOURCE_DIR + "/ngraph_bind",
+                 'ngraph_bind.op': PYNGRAPH_SOURCE_DIR + "/ngraph_bind/op",
+                 'ngraph_bind.op.util': PYNGRAPH_SOURCE_DIR + "/ngraph_bind/op/util",
+                 'ngraph_bind.passes': PYNGRAPH_SOURCE_DIR + "/ngraph_bind/passes",
+                 'ngraph_bind.runtime': PYNGRAPH_SOURCE_DIR + "/ngraph_bind/runtime",
+                 'ngraph': PYNGRAPH_SOURCE_DIR + "/ngraph",
+                 'ngraph.utils': PYNGRAPH_SOURCE_DIR + "/ngraph/utils"},
+    packages = ['ngraph_bind', 'ngraph_bind.op', 'ngraph_bind.op.util', 'ngraph_bind.passes', 'ngraph_bind.runtime', 'ngraph', 'ngraph.utils'],
     cmdclass={'build_ext': BuildExt},
     data_files = data_files,
     install_requires = requirements,
