+++ python/test/ngraph_api/test_ops_binary.py
@@ -18,7 +18,7 @@ import operator
 import numpy as np
 import pytest
 
-import ngraph_api as ng
+import ngraph as ng
 
 
 @pytest.mark.parametrize('ng_api_helper,numpy_function', [
