+++ python/tox.ini
@@ -20,15 +20,15 @@ deps =
   flake8-bugbear
 commands=
   {envbindir}/python setup.py develop
-  flake8 {posargs:pyngraph/ ngraph/ ngraph_api/ examples/}
+  flake8 {posargs:pyngraph/ ngraph/ ngraph_bind/ examples/}
   flake8 --ignore=D100,D101,D102,D103,D104,D105,D107 test/  # ignore lack of docs in tests
-  mypy --config-file=tox.ini {posargs:pyngraph/ ngraph/ ngraph_api/ examples/} # Use MyPy in Python 3 only
+  mypy --config-file=tox.ini {posargs:pyngraph/ ngraph/ ngraph_bind/ examples/} # Use MyPy in Python 3 only
   pytest {posargs:test/}
 
 [testenv:py27]
 commands=
   {envbindir}/python setup.py develop
-  flake8 {posargs:pyngraph/ ngraph/ ngraph_api/ examples/}
+  flake8 {posargs:pyngraph/ ngraph/ ngraph_bind/ examples/}
   flake8 --ignore=D100,D101,D102,D103,D104,D105,D107 test/  # ignore lack of docs in tests
   pytest {posargs:test/}
 
@@ -60,7 +60,7 @@ show_column_numbers = True
 show_none_errors  = True
 
 # put custom per-file options here in sections that map their names into filenames, e.g. gta_workflow/filename.py is
-# [mypy-ngraph_api/filename]
+# [mypy-ngraph/filename]
 
 [mypy-test.*]
 disallow_untyped_defs = False
